3 merge test, 3 swap test(swap fork 저절로 됨), 1 cow test remained, total 7 failed

+++ include/vm/file.h
@@ -11,13 +11,13 @@ struct file_page {
 	void* init_mmaped_va;
     size_t read_bytes;
     size_t zero_bytes;
-    off_t offset;
+    long offset;
     bool writable;
 };
 
 void vm_file_init (void);
 bool file_backed_initializer (struct page *page, enum vm_type type, void *kva);
-void *do_mmap(void *addr, size_t length, int writable,
-		struct file *file, off_t offset);
+void *do_mmap(void *addr, long length, int writable,
+		struct file *file, long offset);
 void do_munmap (void *va);
 #endif
